refactor: content components
