--- src/content/about.mdx
@@ -1,20 +1,0 @@
----
-name: Toshiki Okazaki
-bio: ビール党のエンジニアです。
-avatar: https://github.com/cakegaly.png
-github: cakegaly
----
-
-## 興味
-
-- (Web-)Performance
-- Accessibility (a11y)
-
-## よく使う技術
-
-- React
-- Tailwind CSS
-
-## 座右の銘
-
-- 夙夜夢寐
